Move LibFiles to config module and define a singleton PYP5JS_FILES
