+++ pyp5js/templates_renderers.py
@@ -1,9 +1,9 @@
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 
-from pyp5js.fs import LibFiles, SketchFiles
+from pyp5js.config.fs import PYP5JS_FILES
+from pyp5js.fs import SketchFiles
 
-pyp5js_files = LibFiles()
-templates = Environment(loader=FileSystemLoader(str(pyp5js_files.templates_dir)))
+templates = Environment(loader=FileSystemLoader(str(PYP5JS_FILES.templates_dir)))
 
 
 def get_sketch_index_content(sketch_files):
@@ -25,5 +25,5 @@ def get_target_sketch_content(sketch_files):
     Renders the content to be written in the temporary SKETCH_NAME/target_sketch.py file
     """
     context = {"sketch_name": sketch_files.sketch_name}
-    index_template = templates.get_template(pyp5js_files.target_sketch_template.name)
+    index_template = templates.get_template(PYP5JS_FILES.target_sketch_template.name)
     return index_template.render(context)
