+++ pyp5js/tests/test_templates_renderers.py
@@ -1,11 +1,11 @@
 from pyp5js import templates_renderers as renderers
 from pyp5js.fs import SketchFiles
-
+from pyp5js.config.fs import PYP5JS_FILES
 
 def test_get_sketch_index_content():
     sketch_files = SketchFiles('foo')
 
-    expected_template = renderers.templates.get_template(sketch_files.from_lib.transcrypt_index_html.name)
+    expected_template = renderers.templates.get_template(PYP5JS_FILES.transcrypt_index_html.name)
     expected_content = expected_template.render({
         'sketch_name': sketch_files.sketch_name,
         "p5_js_url": sketch_files.STATIC_NAME + "/p5.js",
@@ -18,7 +18,7 @@ def test_get_sketch_index_content():
 def test_get_target_sketch_content():
     sketch_files = SketchFiles('foo')
 
-    expected_template = renderers.templates.get_template(sketch_files.from_lib.target_sketch_template.name)
+    expected_template = renderers.templates.get_template(PYP5JS_FILES.target_sketch_template.name)
     expected_content = expected_template.render({
         'sketch_name': sketch_files.sketch_name,
     })
